# api/openapi.yaml
openapi: 3.1.0
info:
  title: GetLicense API
  version: 1.0.0
  description: |
    GetLicense is an open-source software licensing platform.
    It provides a REST API for managing products, licenses, machines,
    API keys, and webhooks with multi-tenant isolation and environment separation.

    ## Webhook Delivery

    Outbound webhook deliveries are at-least-once. Every POST sent to a
    customer endpoint includes a JSON envelope and a set of HMAC headers
    that let receivers authenticate the payload and detect duplicates.

    ### Envelope shape

    ```json
    {
      "id":          "<stable event id — equals domain_event_id>",
      "delivery_id": "<per-attempt webhook_event id>",
      "event_type":  "license.created",
      "data":        { ... },
      "timestamp":   "2026-04-24T12:34:56Z"
    }
    ```

    `id` is STABLE across duplicate fanouts (the at-least-once tail of a
    crash-before-checkpoint window) and across operator-initiated
    redeliveries of the same logical event. Receivers MUST dedupe by
    `id`. `delivery_id` is unique per attempt and is useful for log
    correlation when a single logical event is delivered multiple times.

    ### Headers

    The delivery POST sets three headers in addition to `Content-Type:
    application/json`:

    | Header | Meaning |
    | --- | --- |
    | `X-GetLicense-Timestamp` | Unix seconds at signing time (decimal) |
...
    | `X-GetLicense-Delivery-Id` | Per-attempt id (equal to `envelope.delivery_id`) |

...

...

...

...

...
servers:
...
security:
  - BearerAuth: []
...
paths:
...
  /v1/licenses/{id}:
    get:
      operationId: getLicense
      tags: [licenses]
      summary: Get a license
      description: Retrieves a single license by ID.
      parameters:
        - $ref: "#/components/parameters/ResourceID"
      responses:
        "200":
          description: License details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/License"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "429":
          $ref: "#/components/responses/RateLimited"
        "500":
          $ref: "#/components/responses/InternalError"

    patch:
      operationId: updateLicense
      tags: [licenses]
      summary: Partial update of a license
      description: |
        Updates license.overrides or license.expires_at. Both are optional;
        omitted fields are unchanged. Sending overrides replaces the whole
        overrides struct. Sending expires_at=null clears it (makes the
        license perpetual).
      parameters:
        - $ref: "#/components/parameters/ResourceID"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateLicenseRequest"
      responses:
        "200":
          description: Updated license
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/License"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "422":
          $ref: "#/components/responses/ValidationError"

    delete:
      operationId: revokeLicense
      tags: [licenses]
      summary: Revoke a license
      description: Permanently revokes a license. Only active or suspended licenses can be revoked.
      parameters:
        - $ref: "#/components/parameters/ResourceID"
      responses:
        "204":
          description: License revoked
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "422":
          $ref: "#/components/responses/ValidationError"
        "429":
          $ref: "#/components/responses/RateLimited"
        "500":
          $ref: "#/components/responses/InternalError"
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      description: >
        Dual-mode authentication. Pass either a JWT access token (from login)
        or an API key (prefixed gl_live_ or gl_test_) as the Bearer token.
  parameters:
    ResourceID:
      name: id
      in: path
      required: true
      description: Resource UUID
      schema:
        type: string
        format: uuid
...
  responses:
    Unauthorized:
      description: Authentication required or invalid credentials
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    ValidationError:
      description: Validation error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    RateLimited:
      description: Rate limit exceeded
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    InternalError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
  schemas:
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message, doc_url]
          properties:
            code:
              type: string
              description: Machine-readable error code
              examples: ["validation_error"]
            message:
              type: string
              description: Human-readable error description
              examples: ["Invalid product ID"]
            doc_url:
              type: string
              format: uri
              description: Link to error documentation
              examples: ["https://getlicense.io/docs/errors/validation_error"]

    # ── Pagination ───────────────────────────────────────────────────────
    # List endpoints use opaque cursor pagination. Clients should treat
    # next_cursor as a black box — it encodes the (created_at, id) pair
    # of the last row on the current page. Pass it back as the `cursor`
    # query parameter to fetch the next page.
...
    LicenseOverrides:
      type: object
      description: |
        Sparse per-license overrides for quantitative policy fields. Nil
        (omitted) fields mean "inherit from policy". Only quantitative
        fields are overridable — behavioral flags (Floating, Strict,
        ExpirationStrategy, RequireCheckout) are policy-only.
      properties:
        max_machines:
          type: integer
          nullable: true
        max_seats:
          type: integer
          nullable: true
        checkout_interval_sec:
          type: integer
          nullable: true
        max_checkout_duration_sec:
          type: integer
          nullable: true
        validation_ttl_sec:
          type: integer
          nullable: true
          minimum: 60
          maximum: 2592000
          description: |
            Per-license override of policy.validation_ttl_sec. Null inherits
            from policy. Same bounds as the policy field.
...
    License:
      type: object
      required:
        - id
        - account_id
        - product_id
        - policy_id
        - customer_id
        - key_prefix
        - token
        - status
        - environment
        - overrides
        - created_at
        - updated_at
      properties:
        id:
          type: string
          format: uuid
        account_id:
          type: string
          format: uuid
        product_id:
          type: string
          format: uuid
        policy_id:
          type: string
          format: uuid
          description: Policy this license cascades from
        customer_id:
          type: string
          format: uuid
          description: End-user customer who owns this license
        key_prefix:
          type: string
          description: First characters of the license key for identification
        token:
          type: string
          description: Signed license token for offline verification
        status:
          type: string
          enum: [inactive, active, suspended, revoked, expired]
        overrides:
          $ref: "#/components/schemas/LicenseOverrides"
        expires_at:
          type: string
          format: date-time
          nullable: true
          description: Null for perpetual licenses; stamped on first activation if policy uses FROM_FIRST_ACTIVATION basis
        first_activated_at:
          type: string
          format: date-time
          nullable: true
        environment:
          $ref: "#/components/schemas/Environment"
        grant_id:
          type: string
          format: uuid
          nullable: true
          description: |
            Non-null when the license was created by a grantee exercising
            a LICENSE_CREATE capability under a grant. Identifies the
            grant whose constraints the license was issued under.
          examples: ["3b8fab1c-4b2e-4b1a-9f4e-4c6f6f8a9b10"]
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time

    # ── Machine ──────────────────────────────────────────────────────────
...
    Environment:
      type: string
      enum: [live, test]
...
    UpdateLicenseRequest:
      type: object
      description: |
        Partial update. Omitted fields are unchanged. For expires_at, the
        double-pointer distinguishes "unchanged" (omitted) from "explicitly null"
        (setting to null clears expiry, making the license perpetual).
      properties:
        overrides:
          $ref: "#/components/schemas/LicenseOverrides"
        expires_at:
          type: string
          format: date-time
          nullable: true
        customer_id:
          type: string
          format: uuid
          description: Reassign the license to a different customer in the same account.

    # ── Customer schemas (L4) ────────────────────────────────────────────
...