# internal/db/sqlc/gen/models.go
// Code generated by sqlc. DO NOT EDIT.
// versions:
//   sqlc v1.29.0

package sqlcgen

import (
	"net/netip"
	"time"

	"github.com/jackc/pgx/v5/pgtype"
)

type Account struct {
	ID        pgtype.UUID
	Name      string
	Slug      string
	CreatedAt time.Time
}

type AccountMembership struct {
	ID                  pgtype.UUID
	AccountID           pgtype.UUID
	IdentityID          pgtype.UUID
	RoleID              pgtype.UUID
	Status              string
	InvitedByIdentityID pgtype.UUID
	JoinedAt            time.Time
	CreatedAt           time.Time
	UpdatedAt           time.Time
}

type ApiKey struct {
	ID                    pgtype.UUID
	AccountID             pgtype.UUID
	ProductID             pgtype.UUID
	Prefix                string
	KeyHash               string
	Scope                 string
	Label                 *string
	Environment           string
	ExpiresAt             *time.Time
	CreatedAt             time.Time
	LastUsedAt            *time.Time
	LastUsedIp            *netip.Addr
	LastUsedUserAgentHash *string
	CreatedByIdentityID   pgtype.UUID
	CreatedByApiKeyID     pgtype.UUID
	RevokedAt             *time.Time
	RevokedByIdentityID   pgtype.UUID
	RevokedReason         *string
	Permissions           []string
	IpAllowlist           []netip.Prefix
}

type Customer struct {
	ID                 pgtype.UUID
	AccountID          pgtype.UUID
	Email              string
	Name               *string
	Metadata           []byte
	CreatedByAccountID pgtype.UUID
	CreatedAt          time.Time
	UpdatedAt          time.Time
}

type DomainEvent struct {
	ID              pgtype.UUID
	AccountID       pgtype.UUID
	Environment     string
	EventType       string
	ResourceType    string
	ResourceID      *string
	ActingAccountID pgtype.UUID
	IdentityID      pgtype.UUID
	ActorLabel      string
	ActorKind       string
	ApiKeyID        pgtype.UUID
	GrantID         pgtype.UUID
	RequestID       *string
	IpAddress       *netip.Addr
	Payload         []byte
	CreatedAt       time.Time
}

type Entitlement struct {
	ID        pgtype.UUID
	AccountID pgtype.UUID
	Code      string
	Name      string
	Metadata  []byte
	CreatedAt time.Time
	UpdatedAt time.Time
}

type Environment struct {
	ID          pgtype.UUID
	AccountID   pgtype.UUID
	Slug        string
	Name        string
	Description string
	Icon        string
	Color       string
	Position    int32
	CreatedAt   time.Time
	UpdatedAt   time.Time
}

type Grant struct {
	ID               pgtype.UUID
	GrantorAccountID pgtype.UUID
	GranteeAccountID pgtype.UUID
	Status           string
	ProductID        pgtype.UUID
	Capabilities     []string
	Constraints      []byte
	InvitationID     pgtype.UUID
	ExpiresAt        *time.Time
	AcceptedAt       *time.Time
	CreatedAt        time.Time
	UpdatedAt        time.Time
	Label            *string
	Metadata         []byte
}

type Identity struct {
	ID            pgtype.UUID
	Email         string
	PasswordHash  string
	TotpSecretEnc []byte
	TotpEnabledAt *time.Time
	CreatedAt     time.Time
	UpdatedAt     time.Time
}

type IdentitySessionInvalidation struct {
	IdentityID pgtype.UUID
	MinIat     time.Time
	UpdatedAt  time.Time
}

type Invitation struct {
	ID                  pgtype.UUID
	Kind                string
	Email               string
	TokenHash           string
	AccountID           pgtype.UUID
	RoleID              pgtype.UUID
	GrantDraft          []byte
	CreatedByIdentityID pgtype.UUID
	CreatedByAccountID  pgtype.UUID
	ExpiresAt           time.Time
	AcceptedAt          *time.Time
	CreatedAt           time.Time
}

type License struct {
	ID                  pgtype.UUID
	AccountID           pgtype.UUID
	ProductID           pgtype.UUID
	KeyPrefix           string
	KeyHash             string
	Token               string
	Status              string
	ExpiresAt           *time.Time
	CreatedAt           time.Time
	UpdatedAt           time.Time
	Environment         string
	GrantID             pgtype.UUID
	CreatedByAccountID  pgtype.UUID
	CreatedByIdentityID pgtype.UUID
	PolicyID            pgtype.UUID
	Overrides           []byte
	FirstActivatedAt    *time.Time
	CustomerID          pgtype.UUID
}

type LicenseEntitlement struct {
	LicenseID     pgtype.UUID
	EntitlementID pgtype.UUID
	CreatedAt     time.Time
}

type Machine struct {
	ID             pgtype.UUID
	AccountID      pgtype.UUID
	LicenseID      pgtype.UUID
	Fingerprint    string
	Hostname       *string
	Metadata       []byte
	CreatedAt      time.Time
	Environment    string
	LeaseIssuedAt  time.Time
	LeaseExpiresAt time.Time
	LastCheckinAt  time.Time
	Status         string
}

type Policy struct {
	ID                        pgtype.UUID
	AccountID                 pgtype.UUID
	ProductID                 pgtype.UUID
	Name                      string
	IsDefault                 bool
	DurationSeconds           *int32
	ExpirationStrategy        string
	ExpirationBasis           string
	MaxMachines               *int32
	MaxSeats                  *int32
	Floating                  bool
	Strict                    bool
	RequireCheckout           bool
	CheckoutIntervalSec       int32
	MaxCheckoutDurationSec    int32
	ComponentMatchingStrategy string
	Metadata                  []byte
	CreatedAt                 time.Time
	UpdatedAt                 time.Time
	CheckoutGraceSec          int32
	ValidationTtlSec          *int32
}

type PolicyEntitlement struct {
	PolicyID      pgtype.UUID
	EntitlementID pgtype.UUID
	CreatedAt     time.Time
}

type Product struct {
	ID            pgtype.UUID
	AccountID     pgtype.UUID
	Name          string
	Slug          string
	PublicKey     string
	PrivateKeyEnc []byte
	Metadata      []byte
	CreatedAt     time.Time
}

type RecoveryCode struct {
	ID         pgtype.UUID
	IdentityID pgtype.UUID
	CodeHash   string
	CreatedAt  time.Time
	UsedAt     *time.Time
}

type RefreshToken struct {
	ID         pgtype.UUID
	IdentityID pgtype.UUID
	TokenHash  string
	ExpiresAt  time.Time
	CreatedAt  time.Time
}

type RevokedJti struct {
	Jti        pgtype.UUID
	IdentityID pgtype.UUID
	ExpiresAt  time.Time
	Reason     string
	CreatedAt  time.Time
}

type Role struct {
	ID          pgtype.UUID
	AccountID   pgtype.UUID
	Slug        string
	Name        string
	Permissions []string
	CreatedAt   time.Time
	UpdatedAt   time.Time
}

type WebhookDispatcherCheckpoint struct {
	Singleton         bool
	LastDomainEventID pgtype.UUID
	UpdatedAt         time.Time
}

type WebhookEndpoint struct {
	ID                             pgtype.UUID
	AccountID                      pgtype.UUID
	Url                            string
	Events                         []string
	Active                         bool
	CreatedAt                      time.Time
	Environment                    string
	SigningSecretEncrypted         []byte
	PreviousSigningSecretEncrypted []byte
	PreviousSigningSecretExpiresAt *time.Time
}

type WebhookEvent struct {
	ID                    pgtype.UUID
	AccountID             pgtype.UUID
	EndpointID            pgtype.UUID
	EventType             string
	Payload               []byte
	Status                string
	Attempts              int32
	LastAttemptedAt       *time.Time
	ResponseStatus        *int32
	CreatedAt             time.Time
	Environment           string
	DomainEventID         pgtype.UUID
	ResponseBody          *string
	ResponseBodyTruncated bool
	ResponseHeaders       []byte
	NextRetryAt           *time.Time
	ClaimToken            pgtype.UUID
	ClaimExpiresAt        *time.Time
	UpdatedAt             time.Time
}